+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-05-24 09:28:26
-// Commit: fix(storage): fix memory leak by cancelling storage subscription
+// Updated at: 2026-05-24 09:53:29
+// Commit: feat(storage): add bi-directional event bus handler for storage
 
 class storageComponentState {
-  final String commitRef = '1374';
-  final String lastMessage = 'fix(storage): fix memory leak by cancelling storage subscription';
-  final DateTime timestamp = DateTime.parse('2026-05-24T09:28:26+05:00');
+  final String commitRef = '1375';
+  final String lastMessage = 'feat(storage): add bi-directional event bus handler for storage';
+  final DateTime timestamp = DateTime.parse('2026-05-24T09:53:29+05:00');
 }